frontend. factor out duplicate code #84
